Correción de modelo
