+++ lib/teacher/blog/comment.ex
@@ -1,8 +1,7 @@
 defmodule Teacher.Blog.Comment do
   use Ecto.Schema
-  import Ecto.Changeset
-  alias Teacher.Blog.Comment
 
+  import Ecto.Changeset
 
   schema "blog_comments" do
     field :body, :string
@@ -11,10 +10,4 @@ defmodule Teacher.Blog.Comment do
     timestamps()
   end
 
-  @doc false
-  def changeset(%Comment{} = comment, attrs) do
-    comment
-    |> cast(attrs, [:body])
-    |> validate_required([:body])
-  end
 end
