+++ lib/teacher/blog/post.ex
@@ -1,21 +1,14 @@
 defmodule Teacher.Blog.Post do
   use Ecto.Schema
-  import Ecto.Changeset
-  alias Teacher.Blog.Post
 
+  import Ecto.Changeset
 
   schema "blog_posts" do
     field :body, :string
     field :title, :string
-    # tengo que agregar esta linea en el modelo que según es post.ex que esta en model/ pero este es el único post.ex que me genera: 
-    has_many :comments, Teacher.Comment
+
+    has_many :comments, Teacher.Blog.Comment
     timestamps()
   end
 
-  @doc false
-  def changeset(%Post{} = post, attrs) do
-    post
-    |> cast(attrs, [:title, :body])
-    |> validate_required([:title, :body])
-  end
 end
